Payment relationships changed
Direct relationship with invoices and/or debit notes has been abandoned
in favour of AgreementPayment and ActivityPayment. This still allows to
establish which particular invoice and/or debit notes have been covered
by a payment. On the other hand, this model is more flexible (e.g.
allows making partial payments) and makes payment verification easier.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>